Use androidx.preference.Preference (#49)
* Enable MultiDex
* Use PreferenceFragmentCompat

Use PreferenceFragmentCompat instead of Preference.
This will also make the library works with AppCompat and AndroidX
