+++ Sample/build.gradle
@@ -4,12 +4,16 @@ android {
     compileSdkVersion 28
 
     defaultConfig {
-        minSdkVersion 7
+        minSdkVersion 14
         targetSdkVersion 28
+        multiDexEnabled true
     }
 
 }
 
 dependencies {
     implementation project(':ColorPickerPreference')
+    implementation "androidx.appcompat:appcompat:1.1.0"
+    implementation "androidx.preference:preference:1.1.0"
+
 }
