+++ Sample/src/main/java/net/margaritov/preference/colorpicker/Test.java
@@ -17,27 +17,17 @@
 package net.margaritov.preference.colorpicker;
 
 import android.os.Bundle;
-import android.preference.Preference;
-import android.preference.Preference.OnPreferenceChangeListener;
-import android.preference.PreferenceActivity;
 
-public class Test extends PreferenceActivity {
+import androidx.appcompat.app.AppCompatActivity;
+
+public class Test extends AppCompatActivity {
     /**
      * Called when the activity is first created.
      */
     @Override
     public void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        addPreferencesFromResource(R.xml.settings);
-        ((ColorPickerPreference) findPreference("color2")).setOnPreferenceChangeListener(new OnPreferenceChangeListener() {
 
-            @Override
-            public boolean onPreferenceChange(Preference preference, Object newValue) {
-                preference.setSummary(ColorPickerPreference.convertToARGB(Integer.valueOf(String.valueOf(newValue))));
-                return true;
-            }
-
-        });
-        ((ColorPickerPreference) findPreference("color2")).setAlphaSliderEnabled(true);
+        setContentView(R.layout.test);
     }
 }
