+++ build.gradle
@@ -18,5 +18,6 @@ buildscript {
 allprojects {
     repositories {
         jcenter()
+        google()
     }
 }
